fix: move interfaces
