+++ src/stacks/api/report/express-app/interfaces.ts
@@ -1,0 +1,15 @@
+export const REPORT_TABLE_NAME: string = 'REPORT_TABLE_NAME';
+
+export interface AuthorizerContext {
+  user: string;
+}
+
+export interface IReport {
+  timestamp: string;
+  ownerId: string;
+  type?: string;
+}
+
+export interface QueryStringParameters {
+  LastEvaluatedKey?: string;
+}
