+++ src/stacks/api/report/server.ts
@@ -1,0 +1,5 @@
+import app from './express-app';
+
+app.listen(3000, () => {
+  console.log('h');
+});
