+++ src/stacks/api/report/serverless.ts
@@ -4,18 +4,3 @@ import app from './express-app';
 // Wrap the express app with serverless-http
 export const handler = serverless(app);
 
-export const REPORT_TABLE_NAME: string = 'REPORT_TABLE_NAME';
-
-export interface AuthorizerContext {
-  user: string;
-}
-
-export interface IReport {
-  timestamp: string;
-  ownerId: string;
-  type?: string;
-}
-
-export interface QueryStringParameters {
-  LastEvaluatedKey?: string;
-}
